+++ KNUTICEUnitTests/MainNoticesTest/MockMainNoticesViewModelTest.swift
@@ -11,14 +11,14 @@ import RxSwift
 import XCTest
 @testable import KNUTICE
 
-final class MainNoticesViewModelMockTest: XCTestCase {
+final class MockMainNoticesViewModelTest: XCTestCase {
     private var viewModel: MainViewModel!
     private var disposeBag: DisposeBag!
 
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
         let configuration = URLSessionConfiguration.default
-        configuration.protocolClasses = [TopThreeNoticesMockURLProtocol.self]
+        configuration.protocolClasses = [MockTopThreeNoticesURLProtocol.self]
         let session = Session(configuration: configuration)
         Container.shared.remoteDataSource.register {
             RemoteDataSourceImpl(session: session)
